requirements.txt
Django==3.1.2
django-cors-headers==3.5.0
django-rest-swagger==2.2.0
djangorestframework==3.12.1
drf-yasg==1.20.0
mysqlclient==2.0.1
Pillow==8.0.1